+++ src/utils/event.js
@@ -1,0 +1,9 @@
+const preventBubbling = e => {
+  e.stopPropagation();
+};
+
+const preventEvent = e => {
+  e.preventDefault();
+};
+
+export { preventBubbling, preventEvent };
