+++ src/utils/processCard.js
@@ -14,18 +14,4 @@ const makeValidDate = date => {
 const makeCardOwnerName = name =>
   name.length > 10 ? `${name.slice(0, 10)}...` : name;
 
-const preventBubbling = e => {
-  e.stopPropagation();
-};
-
-const preventEvent = e => {
-  e.preventDefault();
-};
-
-export {
-  encryptCardNumber,
-  makeValidDate,
-  makeCardOwnerName,
-  preventBubbling,
-  preventEvent,
-};
+export { encryptCardNumber, makeValidDate, makeCardOwnerName };
